requirements.txt
ccxt==4.2.95
cryptography==41.0.7
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
tensorflow==2.15.0
scikit-learn==1.4.1.post1
matplotlib==3.8.4
mysql-connector-python==8.3.0
tqdm==4.66.2
ta>=0.10.2
seaborn==0.12.2